Refactor CommandResponse struct and RunCmdE function to use more descriptive variable names
The commit refactors the CommandResponse struct and the RunCmdE function to use more descriptive variable names. Specifically, the Out and Err fields of the CommandResponse struct are renamed to Stdout and Stderr, respectively. Similarly, the outBytes and errBytes variables in the RunCmdE function are renamed to stdoutBuffer and stderrBuffer, respectively. This makes the code more readable and easier to understand.