Refactor code for environment variable substitution
This commit includes changes to the `cfg` and `data` packages. In the `cfg` package, a new function called `EnvironmentTemplate` has been added that replaces placeholders in a configuration string with environment variables, if available. The function `GetEnvironmentOrDefault` has been updated to use `zlog` package for logging. In the `data` package, the `CloseRows` function has been removed as it was unused.