Add function NoRows to check if no rows are found in result set
This commit adds a new function called NoRows to the mysql/errors.go file. This function can be used to check if no rows are found in a result set. The function takes an error as input and returns true if there are no rows, and false otherwise. The implementation of the function checks if the input error is nil, and if not, it compares the error string with the value "sql: no rows in result set". If the error string matches, it returns true. Otherwise, it returns false.